chore: move some types to `ProfilingTypes`
